新增回复功能(小部分未完成)

+++ src/main/java/com/nsusoft/community/controller/PublishController.java
@@ -22,7 +22,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id, ModelMap modelMap) {
+    public String edit(@PathVariable(name = "id") Long id, ModelMap modelMap) {
         QuestionDto question = service.queryQustionById(id);
         modelMap.addAttribute("id", question.getId());
         modelMap.addAttribute("title", question.getTitle());
@@ -32,7 +32,7 @@ public class PublishController {
     }
 
     @RequestMapping(value = "/submit", method = RequestMethod.POST)
-    public String doPublish(@RequestParam(value = "id", required = false) Integer id,
+    public String doPublish(@RequestParam(value = "id", required = false) Long id,
                             @RequestParam(value = "title", required = false) String title,
                             @RequestParam(value = "description", required = false) String description,
                             @RequestParam(value = "tag", required = false) String tag,
